move truncate ops to one file
