+++ src/math/operations/truncateOps.ts
@@ -1,9 +1,25 @@
 import Classical from "data/specs/Classical"
-import Operation from "../Operation"
+import Operation from "./Operation"
 import {
   truncate as metaTruncate,
   rectify as metaRectify,
-} from "../../operations-new/truncate"
+} from "../operations-new/truncate"
+
+export const truncate = new Operation<{}, Classical>("truncate", {
+  apply: (solid) => metaTruncate.applyLeft(solid, {}),
+  getResult: ({ specs }) => metaTruncate.getRight(specs),
+  canApplyTo(info): info is Classical {
+    return metaTruncate.canApplyLeftTo(info)
+  },
+})
+
+export const rectify = new Operation<{}, Classical>("rectify", {
+  apply: (solid) => metaRectify.applyLeft(solid, {}),
+  getResult: ({ specs }) => metaRectify.getRight(specs),
+  canApplyTo(info): info is Classical {
+    return metaRectify.canApplyLeftTo(info)
+  },
+})
 
 interface Options {
   facet?: "vertex" | "face"
