--- src/math/operations/truncateOps/index.ts
@@ -1,2 +1,0 @@
-export { truncate, rectify } from "./truncate"
-export { sharpen } from "./sharpen"
